switch back to typeson, but it works now, allow custom encrypt/decrypt methods
