+++ __tests__/options.js
@@ -85,7 +85,7 @@ describe('Options', () => {
         const out = await readingDb.friends.get(1);
         expect(out).toMatchInlineSnapshot(`
             Object {
-              "__encryptedBson": Uint8Array [
+              "__encryptedData": Uint8Array [
                 0,
                 0,
                 0,
@@ -110,109 +110,102 @@ describe('Options', () => {
                 0,
                 0,
                 0,
-                170,
-                216,
-                213,
-                15,
-                220,
-                214,
-                173,
-                130,
-                115,
-                171,
-                87,
-                248,
-                213,
-                164,
-                123,
-                34,
-                47,
-                64,
-                134,
-                13,
-                57,
-                23,
-                63,
-                221,
-                110,
-                96,
-                193,
-                99,
-                26,
-                217,
-                94,
-                218,
-                22,
-                69,
-                184,
-                174,
-                249,
+                81,
+                251,
+                18,
+                148,
                 104,
-                55,
-                148,
-                160,
-                180,
-                189,
-                52,
-                85,
-                203,
-                179,
-                182,
-                178,
-                225,
-                73,
-                185,
-                100,
-                182,
+                242,
+                247,
+                39,
+                168,
+                202,
                 130,
-                95,
-                128,
-                36,
-                1,
-                206,
-                148,
-                149,
-                81,
-                171,
-                89,
-                88,
-                180,
-                44,
-                25,
-                25,
-                193,
-                218,
-                81,
-                187,
-                96,
-                234,
-                218,
-                35,
-                207,
-                116,
-                111,
-                214,
+                130,
+                106,
+                10,
                 38,
-                194,
-                174,
+                121,
+                3,
+                98,
+                232,
+                108,
+                86,
+                28,
+                124,
+                138,
+                41,
                 35,
-                128,
-                222,
+                168,
+                14,
+                115,
+                181,
+                113,
+                218,
+                89,
+                0,
+                246,
+                163,
+                255,
+                13,
+                5,
+                207,
+                245,
+                228,
+                145,
+                15,
+                38,
+                191,
+                193,
+                209,
                 164,
-                91,
-                23,
-                169,
-                189,
-                176,
-                103,
-                120,
-                198,
-                235,
-                134,
-                16,
-                26,
+                225,
+                25,
+                230,
                 35,
-                103,
+                135,
+                227,
+                62,
+                244,
+                4,
+                48,
+                184,
+                207,
+                146,
+                77,
+                171,
+                59,
+                31,
+                252,
+                61,
+                20,
+                77,
+                176,
+                130,
+                1,
+                174,
+                108,
+                253,
+                174,
+                84,
+                205,
+                120,
+                46,
+                152,
+                113,
+                211,
+                170,
+                78,
+                229,
+                178,
+                200,
+                58,
+                90,
+                184,
+                190,
+                190,
+                41,
+                105,
               ],
               "age": 25,
               "id": 1,
@@ -290,7 +283,7 @@ describe('Options', () => {
         const out = await readingDb.friends.get(1);
         expect(out).toMatchInlineSnapshot(`
             Object {
-              "__encryptedBson": Uint8Array [
+              "__encryptedData": Uint8Array [
                 0,
                 0,
                 0,
@@ -315,109 +308,102 @@ describe('Options', () => {
                 0,
                 0,
                 0,
-                170,
-                216,
-                213,
-                15,
-                220,
-                214,
-                173,
-                130,
-                115,
-                171,
-                87,
-                248,
-                213,
-                164,
-                123,
-                34,
-                47,
-                64,
-                134,
-                13,
-                57,
-                23,
-                63,
-                221,
-                110,
-                96,
-                193,
-                99,
-                26,
-                217,
-                94,
-                218,
-                22,
-                69,
-                184,
-                174,
-                249,
+                81,
+                251,
+                18,
+                148,
                 104,
-                55,
-                148,
-                160,
-                180,
-                189,
-                52,
-                85,
-                203,
-                179,
-                182,
-                178,
-                225,
-                73,
-                185,
-                100,
-                182,
+                242,
+                247,
+                39,
+                168,
+                202,
                 130,
-                95,
-                128,
-                36,
-                1,
-                206,
-                148,
-                149,
-                81,
-                171,
-                89,
-                88,
-                180,
-                44,
-                25,
-                25,
-                193,
-                218,
-                81,
-                187,
-                96,
-                234,
-                218,
-                35,
-                207,
-                116,
-                111,
-                214,
+                130,
+                106,
+                10,
                 38,
-                194,
-                174,
+                121,
+                3,
+                98,
+                232,
+                108,
+                86,
+                28,
+                124,
+                138,
+                41,
                 35,
-                128,
-                222,
+                168,
+                14,
+                115,
+                181,
+                113,
+                218,
+                89,
+                0,
+                246,
+                163,
+                255,
+                13,
+                5,
+                207,
+                245,
+                228,
+                145,
+                15,
+                38,
+                191,
+                193,
+                209,
                 164,
-                91,
-                23,
-                169,
-                189,
-                176,
-                103,
-                120,
-                198,
-                235,
-                134,
-                16,
-                26,
+                225,
+                25,
+                230,
                 35,
-                103,
+                135,
+                227,
+                62,
+                244,
+                4,
+                48,
+                184,
+                207,
+                146,
+                77,
+                171,
+                59,
+                31,
+                252,
+                61,
+                20,
+                77,
+                176,
+                130,
+                1,
+                174,
+                108,
+                253,
+                174,
+                84,
+                205,
+                120,
+                46,
+                152,
+                113,
+                211,
+                170,
+                78,
+                229,
+                178,
+                200,
+                58,
+                90,
+                184,
+                190,
+                190,
+                41,
+                105,
               ],
               "age": 25,
               "id": 1,
@@ -493,7 +479,7 @@ describe('Options', () => {
         const out = await readingDb.friends.get(1);
         expect(out).toMatchInlineSnapshot(`
             Object {
-              "__encryptedBson": Uint8Array [
+              "__encryptedData": Uint8Array [
                 0,
                 0,
                 0,
@@ -518,117 +504,109 @@ describe('Options', () => {
                 0,
                 0,
                 0,
-                41,
-                99,
-                28,
-                205,
-                178,
-                15,
-                149,
-                114,
-                204,
-                119,
-                206,
-                135,
-                107,
-                176,
-                210,
-                2,
-                39,
-                64,
-                134,
-                13,
-                57,
-                23,
-                63,
-                221,
-                110,
-                96,
-                193,
-                99,
+                162,
                 26,
-                217,
-                94,
-                218,
-                22,
-                69,
-                184,
-                174,
-                249,
                 104,
-                55,
-                148,
-                160,
-                180,
-                189,
-                52,
-                85,
-                203,
-                179,
-                164,
-                168,
-                241,
-                59,
-                221,
-                1,
-                194,
-                130,
-                79,
                 243,
-                80,
-                115,
-                238,
-                144,
-                146,
-                37,
-                153,
-                104,
-                107,
-                142,
-                29,
-                16,
-                74,
-                230,
-                142,
-                18,
-                237,
-                63,
-                234,
-                178,
-                1,
-                236,
-                105,
-                126,
-                199,
-                54,
-                196,
-                203,
-                33,
-                252,
-                183,
-                199,
-                47,
-                1,
-                186,
-                181,
-                217,
-                7,
-                20,
-                167,
-                197,
-                149,
-                31,
-                16,
-                74,
-                11,
-                202,
-                214,
-                44,
-                245,
-                197,
-                195,
-                56,
+                116,
+                100,
                 222,
+                194,
+                68,
+                123,
+                52,
+                70,
+                113,
+                135,
+                111,
+                42,
+                3,
+                98,
+                232,
+                108,
+                86,
+                28,
+                124,
+                138,
+                41,
+                35,
+                168,
+                14,
+                115,
+                181,
+                113,
+                218,
+                89,
+                0,
+                246,
+                163,
+                255,
+                13,
+                5,
+                207,
+                245,
+                228,
+                145,
+                15,
+                60,
+                175,
+                145,
+                142,
+                240,
+                185,
+                25,
+                175,
+                117,
+                176,
+                231,
+                40,
+                244,
+                6,
+                59,
+                169,
+                176,
+                135,
+                86,
+                255,
+                72,
+                90,
+                189,
+                98,
+                5,
+                81,
+                178,
+                253,
+                87,
+                172,
+                96,
+                251,
+                174,
+                3,
+                147,
+                63,
+                124,
+                203,
+                48,
+                196,
+                190,
+                81,
+                233,
+                252,
+                158,
+                121,
+                23,
+                169,
+                189,
+                176,
+                103,
+                120,
+                198,
+                235,
+                134,
+                16,
+                26,
+                1,
+                26,
               ],
               "age": 25,
               "id": 1,
