+++ __tests__/upgrades.js
@@ -94,7 +94,7 @@ describe('Upgrades', () => {
         const unencrypted = await unencryptedDb.friends.get(1);
         expect(unencrypted).toMatchInlineSnapshot(`
             Object {
-              "__encryptedBson": Uint8Array [
+              "__encryptedData": Uint8Array [
                 0,
                 0,
                 0,
@@ -119,117 +119,109 @@ describe('Upgrades', () => {
                 0,
                 0,
                 0,
-                214,
-                116,
-                214,
-                241,
-                41,
-                81,
-                14,
-                93,
-                54,
-                72,
-                78,
-                240,
-                28,
-                236,
-                145,
-                231,
-                39,
-                64,
-                134,
-                13,
-                57,
-                23,
-                63,
-                221,
-                110,
-                96,
-                193,
-                99,
-                26,
-                217,
-                94,
-                218,
-                22,
-                69,
-                184,
                 174,
-                249,
-                104,
-                55,
-                148,
-                160,
-                180,
-                189,
-                52,
-                85,
-                203,
-                179,
-                164,
-                168,
-                241,
-                59,
-                221,
-                1,
-                194,
-                130,
-                79,
-                240,
-                77,
-                98,
-                255,
-                128,
-                148,
-                64,
-                139,
-                100,
-                107,
-                142,
-                88,
-                18,
-                88,
-                255,
-                199,
-                79,
-                178,
-                100,
-                176,
-                170,
-                79,
-                216,
-                29,
-                14,
-                209,
-                39,
-                194,
-                174,
-                70,
-                248,
-                222,
-                182,
-                91,
-                116,
-                200,
-                149,
-                184,
-                120,
-                96,
-                135,
-                244,
-                197,
-                68,
-                9,
-                75,
-                71,
-                245,
-                195,
-                112,
-                224,
-                206,
                 208,
+                32,
+                66,
+                167,
+                220,
+                65,
+                11,
+                204,
+                195,
+                183,
+                109,
+                113,
+                147,
+                122,
+                117,
+                3,
+                98,
+                232,
+                108,
+                86,
+                28,
+                124,
+                138,
+                41,
+                35,
+                168,
+                14,
+                115,
+                181,
+                113,
+                218,
+                89,
+                0,
+                246,
+                163,
+                255,
+                13,
+                5,
+                207,
+                245,
+                228,
+                145,
+                15,
+                60,
+                175,
+                145,
+                142,
+                240,
+                185,
+                25,
+                172,
+                104,
+                161,
+                246,
                 56,
-                222,
+                242,
+                65,
+                35,
+                177,
+                215,
+                133,
+                68,
+                230,
+                1,
+                7,
+                226,
+                57,
+                95,
+                73,
+                252,
+                201,
+                1,
+                242,
+                39,
+                237,
+                174,
+                83,
+                218,
+                120,
+                120,
+                128,
+                105,
+                146,
+                142,
+                66,
+                255,
+                170,
+                132,
+                106,
+                71,
+                242,
+                164,
+                177,
+                43,
+                71,
+                211,
+                183,
+                147,
+                27,
+                9,
+                1,
+                26,
               ],
               "age": 25,
               "id": 1,
