+++ src/checkForKeyChange.ts
@@ -1,17 +1,25 @@
 import Dexie from 'dexie';
-import { CryptoSettings, CryptoSettingsTable, TableType, CryptoSettingsTableType } from './types';
-import { performDecryption } from './encryptionMethods';
+import {
+    CryptoSettings,
+    CryptoSettingsTable,
+    TableType,
+    CryptoSettingsTableType,
+    EncryptionMethod,
+    DecryptionMethod,
+} from './types';
 
 export function checkForKeyChange<T extends Dexie>(
     db: T,
     oldSettings: TableType<CryptoSettingsTable<T>> | undefined,
     encryptionKey: Uint8Array,
+    encrypt: EncryptionMethod,
+    decrypt: DecryptionMethod,
     onKeyChange: (db: T) => any
 ) {
     try {
         const changeDetectionObj = oldSettings ? oldSettings.keyChangeDetection : null;
         if (changeDetectionObj) {
-            performDecryption(encryptionKey, new Buffer(changeDetectionObj));
+            decrypt(encryptionKey, changeDetectionObj);
         }
     } catch (e) {
         return Dexie.Promise.resolve(onKeyChange(db));
